Create formulae for tsh-ui
tsh-ui is a command line UI that simplifies the usage of tsh tool that is used to connect via teleport https://goteleport.com
This version has capabilities to show the active requests and login with the selected one. In the next versions will add connect to DB or kubernetes cluster.